+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-## 0.1.0-dev1
+## 0.1.0-dev2
 
+* Added a temporary directory to upload file to API 
 * Added a /healthcheck endpoint
 * Added an API endpoint for detectron2 layout detection